Rename codewalker_xml.py to element.py
